extract shared test scaffold
